Add missing g tag
The difference between `SVG_TAGS` array in bel and yo-yoify is the missing g tag in the yo-yoify, which is causing the g tag to be not recognised/rendered in the browser (Chrome), even though its visible in the inspector. I found it was [added to bel in this commit](https://github.com/shama/bel/commit/9fa42e53a00b2558e2d8dc2562f72f356e1dea7e).

Could this list (and some logic) maybe be stored in a shared module of some sort?