Compile already-compiled template strings
This patch allows yo-yoify to also pick up template strings that were
already compiled by an earlier build step, currently either Babel or
Bublé. This way you don't have to disable the template strings transform
when using Babel, which can be a big pain.

Babel compiles template strings in quite a complex way to keep with the
spec. The literal parts of a template string are assigned to a variable
first, and that variable is passed to the compiled template tag function
call. So this patch detects `_taggedTemplateLiteral` helper calls and
takes the template string literal parts from those.