only include helpers if they are needed
this makes the resulting code a bit smaller if some node doesn't use
splat properties or (more rarely) doesn't have children. and if your
codebase doesn't use splats at all, that helper won't be included in the
bundle. otherwise it just saves a few bytes in the `require()` call,
because that wouldn't be eliminated by uglify.